Refactor, resolve race-condition
