+++ src/modules/fileManager/logger.ts
@@ -1,0 +1,13 @@
+import winston from 'winston'
+
+const logger = winston.createLogger({
+  level: 'info',
+  format: winston.format.json(),
+  transports: [
+    new winston.transports.Console({
+      format: winston.format.combine(winston.format.colorize(), winston.format.simple())
+    })
+  ]
+})
+
+export default logger
